requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
grpcio==1.62.0
grpcio-tools==1.62.0
requests==2.31.0